NoRef/OFBIZ patch: content: images: entitymodel: NEW DataResource fields: scpWidth, scpHeight [RENAMED: too likely for name conflicts in existing services!] - these will be needed in future queries (99% certain) for which DataResourceAttribute will simply be too slow.
git-svn-id: svn://data.ilscipio.com/ilscipio/ofbiz_foundation/trunk@17034 55bbc10b-e964-4c8f-a844-a62c6f7d3c80

+++ applications/content/src/org/ofbiz/content/image/ContentImageServices.java
@@ -670,8 +670,8 @@ public abstract class ContentImageServices {
                     
                     // SCIPIO: 2017-08-11: now store width & height in new DataResource fields,
                     // due to very high probability we will need these, and with decent access speed.
-                    dataResource.put("scpWidth", bufNewImg.getWidth());
-                    dataResource.put("scpHeight", bufNewImg.getHeight());
+                    dataResource.put("scpWidth", (long) bufNewImg.getWidth());
+                    dataResource.put("scpHeight", (long) bufNewImg.getHeight());
                     
                     Map<String, Object> customDrFields = new HashMap<>();
                     customDrFields.putAll(ContentImageWorker.RESIZEIMG_DATARESOURCE_FIELDEXPR);
